+++ AgentBridge/Editor/ObjectCommands.cs
@@ -22,7 +22,6 @@ namespace LLMDevTools
         {
             public string    Cmd         => "object_create";
             public string    Description => "Create a new GameObject, optionally as a 3D primitive.";
-            public bool      Core        => true;
             public ArgSpec[] Args        => new[]
             {
                 new ArgSpec("name",      "string", "GameObject", ""),
@@ -67,7 +66,6 @@ namespace LLMDevTools
         {
             public string    Cmd         => "object_delete";
             public string    Description => "";
-            public bool      Core        => true;
             public ArgSpec[] Args        => new[]
             {
                 new ArgSpec("path", "string", "", ""),
@@ -94,7 +92,6 @@ namespace LLMDevTools
         {
             public string    Cmd         => "object_active";
             public string    Description => "";
-            public bool      Core        => true;
             public ArgSpec[] Args        => new[]
             {
                 new ArgSpec("path",   "string", "",     ""),
@@ -124,7 +121,6 @@ namespace LLMDevTools
         {
             public string    Cmd         => "object_rename";
             public string    Description => "";
-            public bool      Core        => true;
             public ArgSpec[] Args        => new[]
             {
                 new ArgSpec("path", "string", "", ""),
@@ -162,7 +158,6 @@ namespace LLMDevTools
         {
             public string    Cmd         => "object_select";
             public string    Description => "Select one or more GameObjects by path.";
-            public bool      Core        => true;
             public ArgSpec[] Args        => new[]
             {
                 new ArgSpec("path",  "string", "", "Single path"),